fix(qov-860): Force uint64 as u64
Despite setting the good format in the openapi, the generated rust type
leads to i32, such as:
```
root_volume_size:
  type: integer
  description: System volume disk size. (in bytes)
  format: uint64
  nullable: true
```

To resolve the issue, force map the u64 type in the openapi generation
command with `--type-mappings=integer+uint64=u64`

+++ src/models/scaleway_baremetal_v1_offer_option_offer_private_network.rs
@@ -15,7 +15,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct ScalewayBaremetalV1OfferOptionOfferPrivateNetwork {
     #[serde(rename = "bandwidth_in_bps", skip_serializing_if = "Option::is_none")]
-    pub bandwidth_in_bps: Option<i32>,
+    pub bandwidth_in_bps: Option<u64>,
 }
 
 impl ScalewayBaremetalV1OfferOptionOfferPrivateNetwork {
