+++ src/models/scaleway_period_baremetal_period_v1_period_schema_period_partition.rs
@@ -20,7 +20,7 @@ pub struct ScalewayPeriodBaremetalPeriodV1PeriodSchemaPeriodPartition {
     pub number: Option<i32>,
     /// (in bytes)
     #[serde(rename = "size", skip_serializing_if = "Option::is_none")]
-    pub size: Option<i32>,
+    pub size: Option<u64>,
     #[serde(
         rename = "use_all_available_space",
         skip_serializing_if = "Option::is_none"
