+++ src/models/scaleway_period_k8s_period_v1_period_cluster_type.rs
@@ -38,7 +38,7 @@ pub struct ScalewayPeriodK8sPeriodV1PeriodClusterType {
     pub resiliency: Option<Resiliency>,
     /// Max RAM allowed for the control plane. (in bytes)
     #[serde(rename = "memory", skip_serializing_if = "Option::is_none")]
-    pub memory: Option<i32>,
+    pub memory: Option<u64>,
     /// Returns information if this offer uses dedicated resources.
     #[serde(rename = "dedicated", skip_serializing_if = "Option::is_none")]
     pub dedicated: Option<bool>,
